my task view and permission unauthorized access functionality done.

+++ routes/web.php
@@ -35,10 +35,10 @@ Route::middleware('guest')->group(function () {
 });
 
 Route::middleware('auth')->group(function () {
-    Route::get('/', [DashboardController::class, 'index'])->name('dashboard');
+    Route::get('/', [DashboardController::class, 'index'])->middleware('permission:dashboard.view')->name('dashboard');
 
-    Route::get('/sub-department/{department}', [AjaxController::class, 'getSubDepartments'])->name('sub_department_by_department');
-    Route::get('/employees-reporting-to', [AjaxController::class, 'getReportingToEmployees'])->name('reporting_to_employees');
+    Route::get('/sub-department/{department}', [AjaxController::class, 'getSubDepartments'])->middleware('permission:employees.add|employees.edit')->name('sub_department_by_department');
+    Route::get('/employees-reporting-to', [AjaxController::class, 'getReportingToEmployees'])->middleware('permission:employees.add|employees.edit')->name('reporting_to_employees');
     Route::get('/show-modal-popup/add-edit', [AjaxController::class, 'getAddEditPopUpForms']);
     Route::get('/show-modal-popup/view', [AjaxController::class, 'getViewPopUpsPage']);
 
@@ -47,15 +47,15 @@ Route::middleware('auth')->group(function () {
     Route::post('/notifications/{notification}/read', [NotificationController::class, 'read'])->name('notifications.read');
     Route::post('/notifications/read-all', [NotificationController::class, 'readAll'])->name('notifications.readAll');
 
-    Route::post('/admin-user/save/{emp_id}', [AdminUserController::class, 'store'])->name('admin_user.save');
-    Route::put('/admin-user/update/{adm_id}', [AdminUserController::class, 'update'])->name('admin_user.update');
+    Route::post('/admin-user/save/{emp_id}', [AdminUserController::class, 'store'])->middleware('permission:admin_users.add')->name('admin_user.save');
+    Route::put('/admin-user/update/{adm_id}', [AdminUserController::class, 'update'])->middleware('permission:admin_users.edit')->name('admin_user.update');
 
-    Route::get('/employees', [EmployeeController::class, 'index'])->name('employees.list');
-    Route::get('employees/add', [EmployeeController::class, 'add'])->name('employees.add');
-    Route::post('employees/save', [EmployeeController::class, 'store'])->name('employees.save');
-    Route::get('employees/view/{emp_id}', [EmployeeController::class, 'view'])->name('employees.view');
-    Route::get('employees/edit/{emp_id}', [EmployeeController::class, 'edit'])->name('employees.edit');
-    Route::put('employees/update/{emp_id}', [EmployeeController::class, 'update'])->name('employees.update');
+    Route::get('/employees', [EmployeeController::class, 'index'])->middleware('permission:employees.view')->name('employees.list');
+    Route::get('employees/add', [EmployeeController::class, 'add'])->middleware('permission:employees.add')->name('employees.add');
+    Route::post('employees/save', [EmployeeController::class, 'store'])->middleware('permission:employees.add')->name('employees.save');
+    Route::get('employees/view/{emp_id}', [EmployeeController::class, 'view'])->middleware('permission:employees.view')->name('employees.view');
+    Route::get('employees/edit/{emp_id}', [EmployeeController::class, 'edit'])->middleware('permission:employees.edit')->name('employees.edit');
+    Route::put('employees/update/{emp_id}', [EmployeeController::class, 'update'])->middleware('permission:employees.edit')->name('employees.update');
 
     Route::get('/my-profile', [MyProfileController::class, 'viewProfile'])->name('my_profile');
     Route::get('/my-profile/edit', [MyProfileController::class, 'editProfile'])->name('my_profile.edit');
@@ -63,30 +63,33 @@ Route::middleware('auth')->group(function () {
     Route::put('/my-profile/update/credentials-info', [MyProfileController::class, 'updateProfileCredentials'])->name('my_profile.update.credentials');
     Route::get('/logout', [AuthenticationController::class, 'logout'])->name('logout');
 
-    Route::get('admin-users', [AdminUserController::class, 'allAdminUsers'])->name('admin_user_module.list');
-    Route::get('admin-user/add', [AdminUserController::class, 'addAdmin'])->name('admin_user_module.add');
-    Route::post('admin-user/store', [AdminUserController::class, 'saveAdmin'])->name('admin_user_module.store');
-    Route::get('admin-user/edit/{adm_id}', [AdminUserController::class, 'editAdmin'])->name('admin_user_module.edit');
-    Route::put('admin-user/update-admin/{adm_id}', [AdminUserController::class, 'updateAdmin'])->name('admin_user_module.update');
-    Route::get('admin-user/view-admin/{adm_id}', [AdminUserController::class, 'viewAdmin'])->name('admin_user_module.view');
+    Route::get('admin-users', [AdminUserController::class, 'allAdminUsers'])->middleware('permission:admin_users.view')->name('admin_user_module.list');
+    Route::get('admin-user/add', [AdminUserController::class, 'addAdmin'])->middleware('permission:admin_user.add')->name('admin_user_module.add');
+    Route::post('admin-user/store', [AdminUserController::class, 'saveAdmin'])->middleware('permission:admin_users.add')->name('admin_user_module.store');
+    Route::get('admin-user/edit/{adm_id}', [AdminUserController::class, 'editAdmin'])->middleware('permission:admin_user.edit')->name('admin_user_module.edit');
+    Route::put('admin-user/update-admin/{adm_id}', [AdminUserController::class, 'updateAdmin'])->middleware('permission:admin_users.edit')->name('admin_user_module.update');
+    Route::get('admin-user/view-admin/{adm_id}', [AdminUserController::class, 'viewAdmin'])->middleware('permission:admin_users.view')->name('admin_user_module.view');
 
-    Route::get('team-members', [TeamMembersController::class, 'index'])->name('team_members.list');
-    Route::get('team-member/view/{called_from}/{member_id}/{pro_id?}', [TeamMembersController::class, 'viewTeamMember'])->name('team_member.view');
+    Route::get('team-members', [TeamMembersController::class, 'index'])->middleware('permission:team_members.view')->name('team_members.list');
+    Route::get('team-member/view/{called_from}/{member_id}/{pro_id?}', [TeamMembersController::class, 'viewTeamMember'])->middleware('permission:team_members.view')->name('team_member.view');
 
-    Route::get('projects/', [ProjectController::class, 'index'])->name('projects.list');
-    Route::get('projects/add', [ProjectController::class, 'addProject'])->name('projects.add');
-    Route::post('projects/save', [ProjectController::class, 'saveProject'])->name('projects.save');
-    Route::get('projects/view/{pro_id}', [ProjectController::class, 'viewProject'])->name('project.view');
+    Route::get('projects/', [ProjectController::class, 'index'])->middleware('permission:projects.view')->name('projects.list');
+    Route::get('projects/add', [ProjectController::class, 'addProject'])->middleware('permission:projects.add')->name('projects.add');
+    Route::post('projects/save', [ProjectController::class, 'saveProject'])->middleware('permission:projects.add')->name('projects.save');
+    Route::get('projects/view/{pro_id}', [ProjectController::class, 'viewProject'])->middleware('permission:projects.view')->name('project.view');
 
 
-    Route::post('task/create', [ProjectTaskController::class, 'store'])->name('task.create');
-    Route::get('all-tasks', [ProjectTaskController::class, 'allTasks'])->name('all_tasks');
-    Route::get('task/view/{prt_id}', [ProjectTaskController::class, 'viewTask'])->name('task.view');
-    Route::post('/update-task-status/{prt_id}', [ProjectTaskController::class, 'updateTaskStatus'])->name('add_task_status_via_ajax');
-    Route::post('/upload-task-attachment/{prt_id}', [ProjectTaskController::class, 'uploadTaskAttachment'])->name('upload_task_attachments_via_ajax');
-    Route::post('/update-sub-task-via-ajax/is-done/{pst_id}', [ProjectSubTaskController::class, 'updateSubTaskIsDone'])->name('update_sub_task.is_done');
-    Route::post('/add-sub-task-via-ajax/{prt_id}', [ProjectSubTaskController::class, 'addSubTask'])->name('add_sub_task_via_ajax');
-    Route::delete('/delete-sub-task-via-ajax/{pst_id}/{prt_id}', [ProjectSubTaskController::class, 'deleteSubTask'])->name('delete_sub_task_via_ajax');
+    Route::post('/task/create', [ProjectTaskController::class, 'store'])->middleware('permission:all_tasks.add')->name('task.create');
+    Route::get('/all-tasks', [ProjectTaskController::class, 'allTasks'])->middleware('permission:all_tasks.view')->name('all_tasks');
+    Route::get('/task/view/{called_from}/{prt_id}', [ProjectTaskController::class, 'viewTask'])->middleware('permission:all_tasks.view|team_tasks.view|my_tasks.view')->name('task.view');
+    Route::post('/update-task-status/{prt_id}', [ProjectTaskController::class, 'updateTaskStatus'])->middleware('permission:all_my_tasks.edit|team_tasks.edit')->name('add_task_status_via_ajax');
+    Route::post('/upload-task-attachment/{prt_id}', [ProjectTaskController::class, 'uploadTaskAttachment'])->middleware('permission:all_my_tasks.edit|team_tasks.edit')->name('upload_task_attachments_via_ajax');
+    Route::post('/update-sub-task-via-ajax/is-done/{pst_id}', [ProjectSubTaskController::class, 'updateSubTaskIsDone'])->middleware('permission:all_my_tasks.edit|team_tasks.edit')->name('update_sub_task.is_done');
+    Route::post('/add-sub-task-via-ajax/{prt_id}', [ProjectSubTaskController::class, 'addSubTask'])->middleware('permission:team_tasks.edit')->name('add_sub_task_via_ajax');
+    Route::delete('/delete-sub-task-via-ajax/{pst_id}/{prt_id}', [ProjectSubTaskController::class, 'deleteSubTask'])->middleware('permission:all_tasks.edit')->name('delete_sub_task_via_ajax');
 
-    Route::get('all-my-tasks', [ProjectTaskController::class, 'assignedToMe'])->name('all_my_tasks');
+    Route::get('all-my-tasks', [ProjectTaskController::class, 'assignedToMe'])->middleware('permission:all_my_tasks.view|all_my_tasks.edit')->name('all_my_tasks');
+    Route::get('all-assigned-tasks', [ProjectTaskController::class, 'assignedByMe'])->middleware('permission:all_assigned_tasks.view|all_assigned_tasks.edit')->name('all_assigned_tasks');
+    Route::get('team-tasks', [ProjectTaskController::class, 'teamTasks'])->middleware('permission:team_tasks.view|team_tasks.edit')->name('all_team_tasks');
+    Route::get('my-tasks', [ProjectTaskController::class, 'myTasks'])->middleware('permission:my_tasks.view|my_tasks.edit')->name('my_tasks');
 });
